collect stack and residual
