+++ src/python/dxl/learn/test/__init__.py
@@ -1,2 +1,3 @@
 from .testcase import TestCase, DistributeTestCase
 from .spy import OpRunSpy
+from .testmodel import UnitBlock
